// src/app/api/swagger/models/api-partisipant-info-for-deal-out-side-e-shop.ts
/* tslint:disable */
/* eslint-disable */
import { ApiOrganizationTypeEnum } from '../models/api-organization-type-enum';
export interface ApiPartisipantInfoForDealOutSideEShopAltDto {
  Address?: string | null;
  BankBik?: string | null;
  BankCheckingAccount?: string | null;
  BankCorrespondentAccount?: string | null;
  BankName?: string | null;
  BankPersonalAccount?: string | null;
  Email?: string | null;
  FirstName?: string | null;
  FullName?: string | null;
  Inn?: string | null;
  IsSMP?: boolean;
  Kpp?: string | null;
  LastName?: string | null;
  MiddleName?: string | null;
  Ogrn?: string | null;
  Ogrnip?: string | null;
  OrganizationType?: ApiOrganizationTypeEnum;
  Phone?: string | null;
  ShortName?: string | null;
  Snils?: string | null;
  WarningMessage?: string | null;
}